refactor: remove unnecessary custom hook
