+++ src/pages/Signin/index.tsx
@@ -1,5 +1,5 @@
-import React, { useEffect } from 'react';
-import { useLocation, useNavigate } from 'react-router-dom';
+import React from 'react';
+import { useNavigate } from 'react-router-dom';
 import Text from '../../components/common/Text';
 import { Logo } from '../../assets/images/landing';
 import {
@@ -8,16 +8,38 @@ import {
   StyledSigninContainer,
   StickyBottom,
 } from './index.styles';
-import { KakaoIcon } from '../../assets/icons';
 import Button, { KakaoButton } from '../../components/common/Button';
-import useSignin from './useSignin';
-import useAuthStore, { AuthLevel } from '../../store/useAuthStore';
-import MainLeftBoxTop from '../../components/common/MainLeftBoxTop';
-import MainLeftBoxBottom from '../../components/common/MainLeftBoxBottom';
+import { AuthLevel } from '../../store/useAuthStore';
+
 import useRequireAuth from '../../hooks/common/useRequireAuth';
 
+const generateUniqueRandomValue = () => {
+  const randomPart = window.crypto
+    .getRandomValues(new Uint32Array(1))[0]
+    .toString(16);
+  const timePart = new Date().getTime().toString(16);
+  return `${randomPart}-${timePart}`;
+};
+
+const signinViaKakao = () => {
+  const state = generateUniqueRandomValue();
+  const nonce = generateUniqueRandomValue();
+
+  sessionStorage.setItem('oauth_state', state);
+  sessionStorage.setItem('openid_nonce', nonce);
+
+  const queryString = new URLSearchParams({
+    client_id: 'de13e7d19e639ae838e4735a0cf14614',
+    redirect_uri: `${process.env.REACT_APP_BASE_URL}/callback`,
+    response_type: 'code',
+    state,
+    nonce,
+  }).toString();
+
+  window.location.href = `https://kauth.kakao.com/oauth/authorize?${queryString}`;
+};
+
 export default function Signin() {
-  const { onKakaoSignin } = useSignin();
   const navigate = useNavigate();
 
   const { isAuthLevelSatisfied } = useRequireAuth(
@@ -55,7 +77,7 @@ export default function Signin() {
               게스트로 입장하기
             </Text>
           </Button>
-          <KakaoButton onClick={onKakaoSignin}>
+          <KakaoButton onClick={signinViaKakao}>
             카카오톡으로 시작하기
           </KakaoButton>
         </StyledButtonContainer>
